Fix and optimize getMsg() so that it works as intended and actually returns text when text exists
I fucked up the DB query in d0fd57d6ce212274e05be0f04331c880ecb2106a.
It would never return anything so the if ( !$row ) condition was always being hit.

Now this actually returns the text it should. As a plus, we got rid of a pointless DB query.

Change-Id: I281d84c9a428949f3336d56eefed0352823007bd